+++ src/ezdxf/enums.py
@@ -1,15 +1,8 @@
 #  Copyright (c) 2021, Manfred Moitzi
 #  License: MIT License
-from enum import IntEnum
+from enum import IntEnum, IntFlag
 from ezdxf.lldxf import const
-from ezdxf.lldxf.const import (
-    TextEntityAlignment,
-    MTextFlowDirection,
-    MTextLineAlignment,
-    MTextStroke,
-    MTextLineSpacing,
-    MTextBackgroundColor,
-)
+from ezdxf.lldxf.const import TextEntityAlignment
 
 
 class MTextEntityAlignment(IntEnum):
@@ -33,3 +26,33 @@ class MTextParagraphAlignment(IntEnum):
     CENTER = 3
     JUSTIFIED = 4
     DISTRIBUTED = 5
+
+
+class MTextFlowDirection(IntEnum):
+    LEFT_TO_RIGHT = const.MTEXT_LEFT_TO_RIGHT
+    TOP_TO_BOTTOM = const.MTEXT_TOP_TO_BOTTOM
+    BY_STYLE = const.MTEXT_BY_STYLE
+
+
+class MTextLineAlignment(IntEnum):  # exclusive state
+    BOTTOM = 0
+    MIDDLE = 1
+    TOP = 2
+
+
+class MTextStroke(IntFlag):  # Combination of flags is possible
+    UNDERLINE = 1
+    STRIKE_THROUGH = 2
+    OVERLINE = 4
+
+
+class MTextLineSpacing(IntEnum):
+    AT_LEAST = const.MTEXT_AT_LEAST
+    EXACT = const.MTEXT_EXACT
+
+
+class MTextBackgroundColor(IntEnum):
+    OFF = const.MTEXT_BG_OFF
+    COLOR = const.MTEXT_BG_COLOR
+    WINDOW = const.MTEXT_BG_WINDOW_COLOR
+    CANVAS = const.MTEXT_BG_CANVAS_COLOR