+++ NEWS.md
@@ -30,6 +30,11 @@ Version 0.17.2b4 - dev
 - DEPRECATED: method `Text.set_align()` will be removed in v1.0.0
 - CHANGE: moved enum `MTextEntityAlignment` to `ezdxf.enums`
 - CHANGE: moved enum `MTextParagraphAlignment` to `ezdxf.enums`
+- CHANGE: moved enum `MTextFlowDirection` to `ezdxf.enums`
+- CHANGE: moved enum `MTextLineAlignment` to `ezdxf.enums`
+- CHANGE: moved enum `MTextStroke` to `ezdxf.enums`
+- CHANGE: moved enum `MTextLineSpacing` to `ezdxf.enums`
+- CHANGE: moved enum `MTextBackgroundColor` to `ezdxf.enums`
 - BUGFIX [#574](https://github.com/mozman/ezdxf/issues/574):
   flattening issues in `Path()` and `ConstructionEllipse()` 
 - BUGFIX: `drawing` add-on shows block references in `ACAD_TABLE` at the